Room notification setting
This patch adds a switch button to the room settings where you can
enable or disable notifications.

When the switch is disabled only mentions are enabled for this room.
When the switch is enabled the setting is set to always notify.

Because this is an on/off button there's no way to fine tuning
notifications, so it's not possible to set never-notify or something
different.

Fix https://gitlab.gnome.org/GNOME/fractal/-/issues/199